Adding other OscRational capabilities (#8)
* Added Rational.jl and updated my test Rational.ipynb

* Added Rational Operator Folder and Conversion.jl

* Added Multiplication.jl and Derivative.jl

* Added Rational.jl and Cauchy.jl

* Removed the notebook version

* Update Rational.jl

* Update Conversion.jl

* Update Derivative.jl

* Update Multiplication.jl

* Update Cauchy.jl

* Update Rational.jl

* Update Domain.jl

Added RationalRealAxis

* Update Dense.jl

Added RationalEvaluationOperator, horner_mat_rat, and Matrix(RatEvalOp)

* Update Conversion.jl

Added rational conversions

* Update Rational.jl

Corrected grid domain to be GridAxis and incorporated oscillatory piece into basis expansion

* Update Conversion.jl

* Update Conversion.jl

Fixed FixedGridValues to have correct RationalEvaluationOperator

* Update ConcreteOperators.jl

Added RationalMatrixOperator

* Update ConcreteOperators.jl

Changed name to BasicMatrixOperator

* Update Derivative.jl

Added correct derivative matrix operator for rational basis without the minus 1

* Update Multiplication.jl

Go back to standard multiplication operator without the -1 included

* Update Conversion.jl

Updated FourierEvaluationOperator to RationalEvaluationOperator

* Update Domain.jl

Fix grid in RationalRealAxis to avoid problems at zero and infinity

* Update Dense.jl

Added alpha as an input to RationalEvaluationOperator and pulled it out of the operator when computing the matrix. Also created diagonal matrix of exp(ikx) to avoid issues with k being a row or column and applying to the right part of the matrix.

* Update ConcreteOperators.jl

Got rid of BasicMatrixOperator since I shouldn't need it to build the derivative operator

* Update Multiplication.jl

Updated * in Rational Multiplication.jl to be on RationalRealAxis rather than PeriodicMappedInterval

* Update Rational.jl

Added alpha as a parameter for the rational basis itself

* Update Multiplication.jl

Added alpha as parameter when we call Rational()

* Update Derivative.jl

Changed function of row and column to return updated elements of derivative matrix for rational expansion derivative without the minus 1 and then built the corresponding banded matrix with that function.

* Update Rational.jl

Added modified grid, and replaced old mdft/midft with my versions of them for the modified grid

* Update Dense.jl

Added DiscreteFourierTransformII that will call the corresponding dft for the rational domain

* Update Conversion.jl

Updated the conversion for the rational basis to use the correct dft

* Update OperatorApproximation.jl

Added Rational and RationalRealAxis as exports to OperatorApproximation.jl

* Update Dense.jl

Added Matrix(DiscreteFourierTransformII)

* Update Basis.jl

Added Rational.jl to Basis.jl so that it's actually included

* Update ConcreteOperators.jl

Added include("Rational/Rational.jl") to ConcreteOperators.jl

* Update Rational.jl

Excluded Cauchy.jl for now since it's causing fatal errors by redefining the M function

* Changed Rational name to OscRational

Changed all "Rational" names to "OscRational" to avoid conflicts with Julia's built in rational type

* Update Domain.jl

Made operations in RationalRealAxis scalar

* Update LazyMatrix.jl

Incorporated Tom's change so that BasicBandedOperator gets built properly

* Update Derivative.jl

Changed how Derivative Operator is being built so that it's correct.

* Fixed alpha issue with Derivatives

Added alpha as an input to OscRational basis and called that in multiplication with Derivative operator so that it could build with the oscillatory parameter.

* Update OscRational.jl

Got rid of alpha dependence in BasisExpansion

* Update Conversion.jl

Added dependence on alpha to RationalEvaluationOperator call

* Update Conversion.jl

Ensuring this is the same as in my tests

* Update OscRational.jl

Fixed size issue in kdft and kidft that accounts for size difference when matrix vs. vector is input

* Adding changes to compute Cauchy integrals using oscillatory rational basis

Updated Cauchy.jl, OscRational.jl (to include Cauchy.jl), and Dense.jl

* Update Cauchy.jl

Ensure that the structure of the operator is correct

* Create test_kaitlynn.jl

Adding my testing file

* C+ is working

Edited DiscreteFourierTransformII matrix to be correct and modified BuildOperatorBlock to have correct structure depending on the sign of alpha. I also added dummy code for the C- case though that isn't working yet

* Update test_kaitlynn.jl

Added all tests for my basis in here

* Adding inner product capability for OscRational and DirectSum

Modified LinAlg.jl to add inner product of DirectSum, modified OperatorApproximation.jl to ensure that the proper functions get exported and overloaded properly, modified Cauchy.jl to include integrals, conjugation, and inner product functionality for OscRational basis expansions, modified Multiplication.jl to be able to multiply two OscRational bases with different \alpha values, modified test_kaitlynn.jl to test all of these capabilities

* Resolved comment in Domain.jl

* Resolved comment in Basis.jl

* Added .ipynb_checkpoints to .gitignore and stopped tracking it

* Resolved comment in OscRational.jl

* Resolved comment in Dense.jl

* Resolved comment with test_kaitlynn.jl

* Resolved issue with Multiplication.jl that is causing tests not to pass

* Resolved issue with last test not passing